requirements.txt
# Application Dependencies
langchain-openai
python-dotenv
